Remove use of language code as country code
Removes the use of http language code as the country code when creating
a new vote. The use of language code as country code was a bad idea for
two reasons. Firstly, the language codes and country codes are not
always the same. Secondly, the language of the voter does not reliably
indicate his nationality. So, it is better to force the voter to choose
his country then to provide a wrong default value.

This fixes the bug that was causing people with Swedish language set in
the browser to have El Salvador as the default country.